moved UserStatus to /main dir added User.interface
