+++ src/app/Context/store.tsx
@@ -1,12 +1,13 @@
 'use client'
 
 import { createContext, useContext, useState } from "react";
-import { UserStatus } from "../main/left-wrapper/enum/temp.enum";
+import { UserStatus } from "../main/enum/UserStatus.enum";
+import { User } from "../main/interface/User.interface";
 
 const globalContext = createContext({});
 
 export const GlobalContextProvider = ({ children }: any) => {
-  const [myInfo, setMyInfo] = useState({
+  const [myInfo, setMyInfo] = useState<User>({
     uid: 98267,
     name: "dohtyleee",
     avatar: "https://cdn.intra.42.fr/users/cde32eb6c2fcfc6871aa7405912c40a7/dohyulee.jpg",
