+++ src/app/main/enum/UserStatus.enum.ts
@@ -1,0 +1,6 @@
+export enum UserStatus {
+  ONLINE = 'ONLINE',
+  OFFLINE = 'OFFLINE',
+  IN_GAME = 'IN_GAME',
+  DEACTIVATED = 'DEACTIVATED',
+}
