+++ src/app/main/left-wrapper/channelItem.tsx
@@ -4,9 +4,9 @@ import React from 'react';
 import styles from './channelItem.module.scss';
 import Image from 'next/image';
 import classnames from "classnames";
-import { UserStatus } from './enum/channel.enum';
 import ChannelItemProps from './interfaces/channelItemProps';
 import DmItemProps from "./interfaces/dmItemProps";
+import { UserStatus } from '../enum/UserStatus.enum';
 
 const ChannelItem = (props: ChannelItemProps) => {
   return (
