+++ src/app/main/left-wrapper/enum/channel.enum.tsx
@@ -1,11 +1,4 @@
-export enum UserStatus {
-    ONLINE = 'ONLINE',
-    OFFLINE = 'OFFLINE',
-    IN_GAME = 'IN_GAME',
-    DEACTIVATED = 'DEACTIVATED',
-  }
-
- export enum ChannelMode {
+export enum ChannelMode {
     PUBLIC = "PUBLIC",
     PRIVATE = "PRIVATE",
     PROTECTED = "PROTECTED",
