+++ src/app/main/left-wrapper/interfaces/dmItemProps.tsx
@@ -1,4 +1,4 @@
-import { UserStatus } from '../enum/channel.enum';
+import { UserStatus } from "../../enum/UserStatus.enum";
 
 interface DmItemProps {
 	friend: string;
